Moved FilePermissionsTest and MetadataDriverTest to their correct namespaces to match their src classes
